+++ lib/filebound_client/endpoints/routes.rb
@@ -22,9 +22,9 @@ module FileboundClient
       # @param [string] notes optional notes to from the user who started the document down the route
       # @return [RoutedItem] a hash of the RoutedItem resource that was added
       def route_document_to_workflow(route_id, document_id, notes)
-        body = { documentId: document_id }
-        body[:notes] if notes
-        put("/routes/#{route_id}", nil, body)
+        params = { documentId: document_id }
+        params[:notes] = notes if notes
+        put("/routes/#{route_id}", params, nil)
       end
 
       # Routes a document to a specified step
@@ -38,14 +38,14 @@ module FileboundClient
       # @param [long] route_step_task_id options RouteStepTask key
       # @return [nil]
       def route_document_to_step(routed_item_id, step_number, comment:, due_date:, user_id:, checklist_data:, route_step_id:, route_step_task_id:)
-        body = {id: routed_item_id, stepNumber: step_number }
-        body[:comment] = comment if comment
-        body[:dueDate] = due_date if due_date
-        body[:userId] = user_id if user_id
-        body[:checklistData] = checklist_data if checklist_data
-        body[:routeStepId] = route_step_id if route_step_id
-        body[:routeStepTaskId] = route_step_task_id if route_step_task_id
-        put('/routes', nil, body)
+        params = {id: routed_item_id, stepNumber: step_number }
+        params[:comment] = comment if comment
+        params[:dueDate] = due_date if due_date
+        params[:userId] = user_id if user_id
+        params[:checklistData] = checklist_data if checklist_data
+        params[:routeStepId] = route_step_id if route_step_id
+        params[:routeStepTaskId] = route_step_task_id if route_step_task_id
+        put('/routes', params, nil)
       end
 
       # Routes a document to a user
@@ -55,8 +55,8 @@ module FileboundClient
       # @param [DateTime] due_date due date for the document
       # @return [nil]
       def route_document_to_user(document_id, user_id, route_back, due_date)
-        body = {documentId: document_id, userId: user_id, routeBack: route_back, dueDate: due_date }
-        put('/routes', nil, body)
+        params = { documentId: document_id, userId: user_id, routeBack: route_back, dueDate: due_date }
+        put('/routes', params, nil)
       end
     end
   end